+++ app/scripts/lib/backbone.layout.js
@@ -5,13 +5,8 @@ define([
 ], function(Backbone, JST) {
 
   Backbone.Layout.configure({
-    manage: true,
     fetchTemplate: function(tmpl) {
       return JST[tmpl];
-    },
-    serialize: function () {
-      var obj = this.collection || this.model;
-      return obj ? obj.toJSON() : {};
     }
   });