+++ app/scripts/models/procs.js
@@ -4,14 +4,26 @@ define([
 ], function (Backbone, URI) {
 
   var Proc = Backbone.Model.extend({
-    idAttribute: 'pid'
+    idAttribute: 'pid',
+
+    initialize: function(attrs, options) {
+      this.listenTo(this.collection, 'sort', this.updateIndex);
+    },
+
+    updateIndex: function() {
+      var idx = this.collection.sortedIndex(this, 'mem');
+
+      this.set('index', idx);
+
+      return this;
+    }
   });
 
   var Procs = Backbone.Collection.extend({
     model: Proc,
 
     comparator: function(p1, p2) {
-      return p1.get('mem') > p2.get('mem') ? -1 : 1;
+      return (p2 && p2.get('mem') < p1.get('mem')) ? 1 : -1
     },
 
     url: function() {
@@ -26,9 +38,10 @@ define([
       this.node = options.node;
       this.num = options.num || 20;
 
-      root.remoteOn('reset', function(models) {
-        root.set(models);
-      });
+      root
+        .remoteOn('reset', function(ms) { root.set(ms); })
+        .remoteOn('add', function(m) { root.add(m); })
+        .remoteOn('remove', function(m) { root.remove(m); });
     }
 
   });