+++ app/scripts/views/nodelist.js
@@ -15,13 +15,7 @@ define([
     initialize: function (options) {
       this.collection.fetch();
       this.collection.on('sync', this.render, this);
-    },
-
-    onItemClick: function (e) {
-      var name = $(e.target).data('name');
-      this.model.selectNode(name);
     }
-
   });
 
   return View;